requirements.txt
flask>=3.0.0
python-telegram-bot>=20.7
requests>=2.31.0
python-dotenv>=1.0.0
cryptography>=41.0.0
chardet>=5.0.0
